requirements.txt
actionlib==1.13.2
albumentations==1.1.0
angles==1.9.13
base-local-planner==1.17.1
bondpy==1.8.6
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.15.0
cycler==0.11.0
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
dynamic-reconfigure==1.7.1
dynamixel-sdk==3.7.51
einops==0.3.2
fonttools==4.28.1
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
image-geometry==1.15.0
imageio==2.10.5
interactive-markers==1.12.0
joblib==1.1.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
kiwisolver==1.3.2
laser-geometry==1.6.7
matplotlib==3.5.0
message-filters==1.15.13
monai==0.8.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pyparsing==2.2.2
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qudida==0.0.4
resource-retriever==1.12.6
rosbag==1.15.13
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.13
roslaunch==1.15.13
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.13
rosmake==1.15.8
rosmaster==1.15.13
rosmsg==1.15.13
rosnode==1.15.13
rosparam==1.15.13
rospy==1.15.13
rosserial-python==0.9.2
rosservice==1.15.13
rostest==1.15.13
rostopic==1.15.13
rosunit==1.15.8
roswtf==1.15.13
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.12
rqt-web==0.4.10
rviz==1.14.10
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
sensor-msgs==1.13.1
setuptools-scm==6.3.2
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
topic-tools==1.15.13
torch==1.10.0
torchensemble==0.1.6
torchvision==0.11.1
tqdm==4.62.3
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
typing-extensions==4.0.0
xacro==1.14.10
